+++ src/main/java/com/cvent/pangaea/ExampleConfiguration.java
@@ -2,50 +2,8 @@ package com.cvent.pangaea;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
-//
-//default_multi_env_config: &default_env_config
-//    database: &default_db
-//        checkConnectionWhileIdle: true
-//        driverClass: "com.microsoft.sqlserver.jdbc.SQLServerDriver"
-//        initialSize: 5
-//        maxConnectionAge: "10s"
-//        minSize: 5
-//        maxSize: 5
-//        validationQuery: "SELECT 1"
-//        isLazy: true
-//    surveyUrlAppRoot: "/Surveys"
-//
-//environmentConfig:
-//    S115:
-//        <<: *default_env_config
-//        database:
-//            <<: *default_db
-//            user: "cvent"
-//            password: "n0rth"
-//            url: "jdbc:sqlserver://a1-dba-115.a1.cvent.com\\dev_silo:50000;database=CVENT_PROD"
-//        reportDatabase:
-//            <<: *default_db
-//            user: "cvent"
-//            password: "n0rth"
-//            url: "jdbc:sqlserver://a1-dba-115.a1.cvent.com\\dev_silo:50000;database=CVENT_REPORT"
-//        surveyUrlDefaultDomain: "silo115-guest.a1.cvent.com"
-//        customFieldServiceDomain: "http://staging-wiz-01.cvent.net:8122/customfields"
-//        defaultEnvironmentConfiguration: true
-//    S116:
-//        <<: *default_env_config
-//        database:
-//            <<: *default_db
-//            user: "cvent"
-//            password: "n0rth"
-//            url: "jdbc:sqlserver://a1-dba-116.a1.cvent.com\\dev_silo:50000;database=CVENT_PROD"
-//        reportDatabase:
-//            <<: *default_db
-//            user: "cvent"
-//            password: "n0rth"
-//            url: "jdbc:sqlserver://a1-dba-116.a1.cvent.com\\dev_silo:50000;database=CVENT_REPORT"
-//        surveyUrlDefaultDomain: "silo116-guest.a1.cvent.com"
-//        customFieldServiceDomain: "http://staging-wiz-01.cvent.net:8122/customfields"
 /**
  * This is just an example configuration object that is meant to be used as part of the dropwizard configuration
  * hierarchy.
@@ -64,13 +22,13 @@ class ExampleConfiguration {
     @JsonIgnore
     public MultiEnvAware<String> getSurveyUrlDefaultDomains() {
         return environmentConfig.convert((env, conf) -> conf.getSurveyUrlDefaultDomain(),
-                new SiloTemplateResolver<>(String.class));
+                new SiloTemplateResolver<>(String.class, new ObjectMapper()));
     }
 
     @JsonIgnore
     public MultiEnvAware<String> getSurveyUrlAppRoots() {
         return environmentConfig.convert((env, conf) -> conf.getSurveyUrlAppRoot(),
-                new SiloTemplateResolver<>(String.class));
+                new SiloTemplateResolver<>(String.class, new ObjectMapper()));
     }
 
     @JsonIgnore
@@ -83,7 +41,8 @@ class ExampleConfiguration {
     }
 
     public MultiEnvAware<MultiEnvConfig> getEnvironmentConfig() {
-        return environmentConfig.convert((env, conf) -> conf, new SiloTemplateResolver<>(MultiEnvConfig.class));
+        return environmentConfig.convert((env, conf) -> conf, new SiloTemplateResolver<>(MultiEnvConfig.class, 
+            new ObjectMapper()));
     }
 
     /**